changed torch.bfloat16 to torch.float32 if we use mps, as mps seems to require it. Fixes #35

+++ src/autoforge/auto_forge.py
@@ -442,7 +442,7 @@ def start(args):
     )
 
     with torch.no_grad():
-        with torch.autocast(device.type, dtype=torch.bfloat16):
+        with torch.autocast(device.type, dtype=dtype):
             if args.perform_pruning:
                 optimizer.prune(
                     max_colors_allowed=args.pruning_max_colors,
